requirements.txt
PyMuPDF==1.24.0
python-docx==0.8.11
python-pptx==0.6.21
transformers==4.36.2
torch
streamlit==1.28.0
pyyaml==6.0.1
pyobjc-framework-FSEvents>=9.2
pyobjc-framework-Cocoa>=9.2
pandas==2.2.0
openpyxl==3.1.2 